Rename user-features-lib
